% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_templates.R
\docType{data}
\name{df_templates}
\alias{df_templates}
\title{Metadata of templates}
\format{
\subsection{\code{df_templates} A data frame with 6 rows and 11 columns containing}{

template metadata, including file paths, names, labels, temporal bounds,
sample rates, frequency ranges, and spectrogram parameters (window length
and overlap).
}
}
\usage{
data(df_templates)
}
\description{
A data frame containing the metadata of the templates to run the examples of
the \code{monitoraSom} package, after importing them back to the R session with
the \code{fetch_template_metadata()} function.
}
\keyword{datasets}
