% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_validations_n.R
\name{diagnostic_validations_n}
\alias{diagnostic_validations_n}
\title{Compute detection diagnostics of validated detections from multiple templates}
\usage{
diagnostic_validations_n(val_n, diag_method = "Auto", pos_prob = 0.95)
}
\arguments{
\item{val_n}{A tibble containing the validation results of the detections
`validate_by_overlap()` function or within the validation app.}

\item{diag_method}{A character string indicating the method to use for
diagnostic validations. The two methods available are: "Auto" (default) or
"Manual". If "Auto" is selected, the function will automatically determine
the cutpoint to be used in the diagnostic validations. If "Manual" is
selected, the user must provide the cutpoint to be used in the diagnostic
validations.}

\item{pos_prob}{A numeric value indicating the probability of a positive test}
}
\value{
A list of results from the diagnostic validations, or NULL if all validations failed.
}
\description{
#' @description `r lifecycle::badge("experimental")`
}
\details{
This function performs diagnostic validations on the detections from
`diagnostic_validations_i()` in cases where there are detections generated
for multiple templates or classes. The diagnostic validations are used to
determine the optimal cutpoint to be used in the detections. The function
will automatically determine the cutpoint to be used in the diagnostic
validations if the "Auto" method is selected. If the "Manual" method is
selected, the user must provide the cutpoint to be used in the diagnostic
validations.
}
