% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_matching_helpers.R
\name{.assert_flim_within_nyquist}
\alias{.assert_flim_within_nyquist}
\title{flim (kHz) must not exceed the file's Nyquist (RMC-11)}
\usage{
.assert_flim_within_nyquist(samp_rate, flim, which_file)
}
\arguments{
\item{samp_rate}{sample rate in Hz.}

\item{flim}{frequency limits \code{c(min, max)} in kHz.}

\item{which_file}{label used in the assertion message.}
}
\description{
flim (kHz) must not exceed the file's Nyquist (RMC-11)
}
