% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_matching_helpers.R
\name{.assert_match_matrices}
\alias{.assert_match_matrices}
\title{Soundscape/template matrices must share bins and the soundscape be longer
(RMC-11 equal frequency resolution, RMC-13 length).}
\usage{
.assert_match_matrices(mat_ss, mat_tp)
}
\arguments{
\item{mat_ss}{soundscape spectrogram matrix.}

\item{mat_tp}{template spectrogram matrix.}
}
\description{
Soundscape/template matrices must share bins and the soundscape be longer
(RMC-11 equal frequency resolution, RMC-13 length).
}
