% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_detections.R
\name{.schema_detections}
\alias{.schema_detections}
\title{Empty (or n-row NA) detections frame in standard schema/order.}
\usage{
.schema_detections(n = 0L)
}
\arguments{
\item{n}{number of NA-filled rows to create (default 0).}
}
\description{
Empty (or n-row NA) detections frame in standard schema/order.
}
