% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_signals.R
\name{.schema_signals}
\alias{.schema_signals}
\title{Empty (or n-row NA) signals frame in standard schema/order.}
\usage{
.schema_signals(n = 0L)
}
\arguments{
\item{n}{number of NA-filled rows to create (default 0).}
}
\description{
Empty (or n-row NA) signals frame in standard schema/order.
}
