% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_external_detections.R
\name{import_external_detections}
\alias{import_external_detections}
\title{Import detections made in BirdNET or other third-party software}
\usage{
import_external_detections(input_path, names_vec = NA, output_path = NA)
}
\arguments{
\item{input_path}{Path to the CSV file containing the detections.}

\item{names_vec}{A named character vector to rename the columns of the #'
input. The default values are: `c("template_name" = "scientific_name",
"detection_start" = "start_time", "detection_end" = "end_time",
"peak_score" = "confidence", "soundscape_path" = "filepath",
"detec_min_score" = "min_conf", "bn_lat" = "lat", "bn_lon" = "lon",
"bn_sensitivity" = "sensitivity", "bn_model" = "model_name")`}

\item{output_path}{Path to save the adapted detection table. If not provided,
the function will return the adapted table.}
}
\value{
A detection table adapted to the `monitoraSom` workflow
}
\description{
`r lifecycle::badge("experimental")`
}
