% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_templates.R
\docType{data}
\name{ls_templates}
\alias{ls_templates}
\title{List of templates}
\format{
\subsection{\code{ls_templates} A list of six \code{tuneR::Wave} objects containing}{

templates to run the examples of the \code{monitoraSom} package. All templates
contain signals corresponding to the last two notes of Basileuterus
culicivorus' song. Names of the list elements are the names of the \code{.wav}
files, and correspond to the name code for templates in the \code{monitoraSom}
workflow.
}
}
\usage{
data(ls_templates)
}
\description{
A list of templates to run the examples of the \code{monitoraSom} package.
}
\keyword{datasets}
