% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segmentation_preset.R
\name{make_segmentation_preset}
\alias{make_segmentation_preset}
\title{Function to create segmentation presets for soundscape analysis}
\usage{
make_segmentation_preset(
  preset_path = NULL,
  preset_id = NULL,
  validation_user,
  soundscapes_path,
  roi_tables_path,
  cuts_path,
  fastdisp = TRUE,
  label_angle = 90,
  show_label = TRUE,
  dyn_range = c(-60, 0),
  wl = 1024,
  ovlp = 0,
  color_scale = "inferno",
  wav_player_type = "R session",
  wav_player_path = "play",
  session_notes = NULL,
  zoom_freq = c(0, 10),
  nav_autosave = FALSE,
  sp_list = "CBRO-2021 (Brazil)"
)
}
\arguments{
\item{preset_path}{Path to save the preset file (optional)}

\item{preset_id}{Identification for the preset (optional)}

\item{validation_user}{User responsible for the validation of the
segmentation}

\item{soundscapes_path}{Path to soundscapes files}

\item{roi_tables_path}{Path to region of interest tables files}

\item{cuts_path}{Path to cuts files}

\item{fastdisp}{If TRUE, the next cut is automatically displayed when the
user validates a cut}

\item{label_angle}{Angle to draw the cut labels. Must be between 0 and 180,
and multiple of 10}

\item{show_label}{If TRUE, the cut labels will be displayed}

\item{dyn_range}{Dynamic range. A vector with two numeric values. Must be
between -100 and 10, and multiple of 10}

\item{wl}{Window length. Must be one of the alternatives: 128, 256, 512,
1024, 2048, or 4096}

\item{ovlp}{Overlap. Must be a numeric value, between 0 and 80, and multiple
of 10}

\item{color_scale}{The color scale for the spectrogram}

\item{wav_player_type}{The type of wav player. "R session" for R
session-based player, "system" for system player}

\item{wav_player_path}{Path to the wav player executable (only for system
player)}

\item{session_notes}{Notes related to the analysis session}

\item{zoom_freq}{Vector with two numeric values, indicating the frequency
range to zoom in the spectrogram}

\item{nav_autosave}{If TRUE, the current segmentation is saved when the user
navigates to another file}

\item{sp_list}{Species list used in the analysis}
}
\value{
A list with the segmentation preset parameters
}
\description{
`r lifecycle::badge("experimental")`

This function creates a list with the segmentation preset parameters for
soundscape analysis, such as soundscapes path, region of interest tables
path, cuts path, fast display, label angle, and dynamic range, among others.
The validation user parameter is mandatory. If the 'soundscapes_path'
parameter is not found locally, a warning message will be shown, and if the
roi_tables_path or cuts_path do not exist, the function will create new
folders.
}
