% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_validation_preset.R
\name{make_validation_preset}
\alias{make_validation_preset}
\title{Creates a validation preset for a soundscape dataset.}
\usage{
make_validation_preset(
  preset_path = NULL,
  preset_name = NULL,
  validation_user,
  templates_path,
  soundscapes_path,
  input_path,
  output_path,
  wav_cuts_path,
  spec_path,
  diag_tab_path,
  wav_player_path = "play",
  wav_player_type = "HTML player",
  val_subset = c("NA", "TP", "FP", "UN"),
  min_score = 0,
  time_pads = 1,
  ovlp = 0,
  wl = 2048,
  dyn_range = c(-60, 0),
  color_scale = "inferno",
  zoom_freq = c(0, 4),
  nav_shuffle = FALSE,
  seed = 123,
  auto_next = FALSE,
  nav_autosave = FALSE,
  overwrite = FALSE,
  session_notes
)
}
\arguments{
\item{preset_path}{Path to save the validation preset. Default is NULL.}

\item{preset_name}{Name for the validation preset. Default is NULL.}

\item{validation_user}{User name for the validation session.}

\item{templates_path}{Path to the directory containing the templates.}

\item{soundscapes_path}{Path to the directory containing the soundscapes.}

\item{input_path}{Path to the input file.}

\item{output_path}{Path to the output file.}

\item{wav_cuts_path}{Path to the directory where the waveform cuts will be saved.}

\item{spec_path}{Path to the directory where the spectrograms will be saved.}

\item{diag_tab_path}{Path to the directory where the diagnostic tables will be saved.}

\item{wav_player_path}{Path to the executable of the waveform player. Default is "play".}

\item{wav_player_type}{Type of waveform player. Default is "HTML player".}

\item{val_subset}{Vector with the subset of the validation session to be performed.}

\item{min_score}{Minimum score required to accept a detection. Default is 0.}

\item{time_pads}{Time padding in seconds for waveform cuts. Default is 1.}

\item{ovlp}{Overlap in percentage for waveform cuts. Default is 0.}

\item{wl}{Window length in samples for spectrograms. Default is 2048.}

\item{dyn_range}{Range in decibels for spectrograms. Default is c(-60, 0).}

\item{color_scale}{Color scale for spectrograms. Default is "inferno".}

\item{zoom_freq}{Frequency range for spectrogram zooming. Default is c(0, 4).}

\item{nav_shuffle}{Shuffle navigation. Default is FALSE.}

\item{seed}{Random seed. Default is 123.}

\item{auto_next}{Automatic transition to the next detection. Default is FALSE.}

\item{nav_autosave}{Autosave navigation. Default is FALSE.}

\item{overwrite}{Overwrite output file. Default is FALSE.}

\item{session_notes}{Notes about the validation session.}
}
\value{
A list with validation session settings.
}
\description{
make_validation_preset
}
\examples{
make_validation_preset(validation_user = "user", templates_path = "templates", soundscapes_path = "soundscapes", input_path = "input", output_path = "output")

}
