% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_n.R
\name{match_n}
\alias{match_n}
\title{Batch template matching}
\usage{
match_n(
  df_grid,
  score_method = "cor",
  ncores = 1,
  output = "detections",
  output_file = NULL,
  autosave_action = "append",
  buffer_size = "template",
  min_score = NULL,
  min_quant = NULL,
  top_n = NULL
)
}
\arguments{
\item{df_grid}{The output of the function 'fetch_match_grid()}

\item{score_method}{A character string indicating the method to use for
matching. The two methods available are: "cor" (Pearson correlation
coefficient) or "dtw" (dynamic time warping). It defaults to "cor".}

\item{ncores}{An integer indicating the number of cores to be used for
parallelization. It defaults to 1.}

\item{output}{A character string indicating the output of the function. The
two options are: "detections" or "scores". If "detections" is
selected, the function will return the output of the function
'fetch_score_peaks_i()'. If "scores" is selected, the function will return
the raw output of the matching algorithm. It defaults to "detections".}

\item{output_file}{A character string indicating the path to save the results
in the format of a CSV file for detections or RDS file for scores. Default
is NULL. We recommend to export detection or raw score files to the
"detections/" subdirectory.}

\item{autosave_action}{A character string indicating the action to be taken
if the output file already exists. Possible values are "append" and
"overwrite". To avoid overwriting existing files, set to "append", but be
aware that it can result in duplicated entries in the output file if the
function is run again. It defaults to "append".}

\item{buffer_size}{A character string indicating the size of the buffer to be
used in the function 'fetch_score_peaks_i()'. The two options are:
"template" or the number of frames of the template spectrogram to
be used as buffer. It defaults to "template".}

\item{min_score}{A numeric value indicating the minimum score to be used in
the function 'fetch_score_peaks_i()'. It defaults to NULL.}

\item{min_quant}{A numeric value indicating the minimum quantile to be used
in the function 'fetch_score_peaks_i()'. It defaults to NULL.}

\item{top_n}{A numeric value indicating the number of top detections to be
used in the function 'fetch_score_peaks_i()'. It defaults to NULL.}
}
\value{
If the format is set to scores, a tibble containing the input data
  frame with an additional column "score_vec". This column contains dataframes
  with "time_vec" (time of each frame) and "score_vec" (matching score at
  that frame) columns. The scores are padded at the start and end to match the
  soundscape spectrogram length. If the format is set to detections, the
  function returns a data frame with the detections. If no paths are provided,
  the output are returned to the R session.
}
\description{
`r lifecycle::badge("experimental")`

  This function is wrapper of 'match_i()' to perform a batch computation of
  the matching vectors of the template and soundscape matches defined in the
  metadata grid obtained as output of the function 'fetch_match_grid()'. The
  available algorithms to compare spectrograms and compute matching scores
  are the Pearson correlation coefficient ("cor") or dynamic time warping
  ("dtw").
}
