% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_match_i.R
\name{plot_match_i}
\alias{plot_match_i}
\title{plot_match_i - Plots the results of the match_i function}
\usage{
plot_match_i(
  tpmatch_detecs_i,
  corr_cut = 0.3,
  wl = 1024,
  flim = c(0, 12),
  dyn.range = c(-80, -10),
  ovlp = 50,
  ...
)
}
\arguments{
\item{tpmatch_detecs_i}{The output of the function tpmatch::match_i}

\item{corr_cut}{The correlation threshold above which to plot detected templates}

\item{wl}{The length of the window (in samples) used for the STFT calculation}

\item{flim}{A vector of length 2 specifying the frequency limits (in kHz) of the plot}

\item{dyn.range}{A vector of length 2 specifying the dynamic range (in dB) of the plot}

\item{ovlp}{The overlap (in %) between adjacent windows used for the STFT calculation}

\item{...}{Other arguments passed to the sonograma function from the seewave package}
}
\value{
A plot of the spectrogram with detected templates highlighted
}
\description{
plot_match_i - Plots the results of the match_i function
}
\examples{
plot_match_i(tpmatch_detecs_i, corr_cut = 0.4, wl = 2048, flim = c(0, 20),
              dyn.range = c(-100, -20), ovlp = 75)
}
