% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_by_overlap_i.R
\name{validate_by_overlap_i}
\alias{validate_by_overlap_i}
\title{Validate detections by overlap with Regions of Interest (ROIs)}
\usage{
validate_by_overlap_i(df_rois, df_detecs, det_species)
}
\arguments{
\item{df_rois}{A data frame containing Regions of Interest (ROIs) as in the
output of `fetch_rois_n()`.}

\item{df_detecs}{A data frame containing detections as in the output of
`fetch_score_peaks_n()`.}

\item{det_species}{A character string specifying the species to validate.}
}
\value{
A data frame with the validated detections.
}
\description{
`r lifecycle::badge("experimental")`

This function takes in two data frames, one containing Regions of Interest
(ROIs) and another containing detections. It then validates the detections by
checking for overlaps with the ROIs. The function returns a data frame with
the validated detections.
}
