% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_by_overlap_n.R
\name{validate_by_overlap_n}
\alias{validate_by_overlap_n}
\title{Batch validation of detections by overlap with Regions of Interest (ROIs)}
\usage{
validate_by_overlap_n(df_rois, df_detecs, validation_user)
}
\arguments{
\item{df_rois}{A data frame containing Regions of Interest (ROIs) as in the
output of `fetch_rois_n()`.}

\item{df_detecs}{A data frame containing detections as in the output of
`fetch_score_peaks_n()`.}

\item{validation_user}{A character string specifying the user name.}
}
\value{
A data frame with the validated detections.
}
\description{
`r lifecycle::badge("experimental")`

  This function takes in two data frames, one containing Regions of Interest
  (ROIs) and another containing detections. It then validates the detections
  by checking for overlaps with the ROIs. The function returns a data frame
  with the validated detections.
}
