% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecs_to_rois.R
\name{detecs_to_rois}
\alias{detecs_to_rois}
\title{Convert detections to ROIs}
\usage{
detecs_to_rois(
  df_detecs,
  username = NULL,
  output_path = NULL,
  filter_tp = FALSE
)
}
\arguments{
\item{df_detecs}{A data frame containing detections or path to a CSV file
with detections in the 'monitoraSom' format.}

\item{username}{A character string identifying the user creating the ROIs.}

\item{output_path}{Path where the ROIs will be saved. We avise against
providing the same path as the one used to export ROI tables obtained from
the segmentation app.}

\item{filter_tp}{Logical indicating whether to filter the detections to only
include true positives. Defaults to FALSE.}
}
\value{
A data frame containing the ROI information.
}
\description{
`r lifecycle::badge("experimental")`

  This function converts detections from template matching or external
  sources into ROIs (Regions of Interest) in the 'monitoraSom' format.
}
\examples{
# Load example detections data
data(df_detecs)

# Convert detections to ROIs
rois <- detecs_to_rois(
  df_detecs = df_detecs,
  username = "example_user"
)

}
