% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecs_to_rois.R
\name{detecs_to_rois}
\alias{detecs_to_rois}
\title{Convert detections to ROIs}
\usage{
detecs_to_rois(
  df_detecs,
  username = NULL,
  output_path = NULL,
  filter_tp = FALSE
)
}
\arguments{
\item{df_detecs}{A data frame containing detections or path to a CSV file
with detections in the 'monitoraSom' format.}

\item{username}{A character string identifying the user creating the ROIs.}

\item{output_path}{Path where the ROIs will be saved. We avise against
providing the same path as the one used to export ROI tables obtained from
the segmentation app.}

\item{filter_tp}{Logical indicating whether to filter the detections to only
include true positives. Defaults to FALSE.}
}
\value{
A data frame containing the ROI information.
}
\description{
`r lifecycle::badge("experimental")`

  This function converts detections from template matching or external
  sources into ROIs (Regions of Interest) in the 'monitoraSom' format.
}
\examples{
\dontrun{

# Load the necessary packages to run this example
library(monitoraSom)
library(dplyr)

# Load a dataset with validated detections (output of the of validation app)
data(df_detecs_val_manual)
glimpse(df_detecs_val_manual)

# Create a directory to store the roi tables
roi_cuts_path <- "./130_rois_from_detections"
dir.create(roi_cuts_path)

# Convert the detections to rois to a unified roi table within the current R
# session and export it as multiple roi tables in the directory created earlier
df_detecs_to_rois <- detecs_to_rois(
    df_detecs = df_detecs_val_manual, username = "Rosa G.L.M.",
    output_path = roi_cuts_path
)

# Check the object
glimpse(df_detecs_to_rois)

# Check the roi tables in the directory created earlier
list.files(roi_cuts_path, pattern = "*.csv")
}

}
