% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_to_rois.R
\name{detecs_to_rois}
\alias{detecs_to_rois}
\title{Convert detections to ROIs}
\usage{
detecs_to_rois(input = df_detecs, output = NULL, username = NULL)
}
\arguments{
\item{input}{A data frame containing detections or path to a CSV file with
detections in the 'monitoraSom' format.}

\item{output}{Path where the ROIs will be saved. If NULL, ROIs will be saved
in "./roi_from_detecs/" in the current working directory.}

\item{username}{A character string identifying the user creating the ROIs.}
}
\value{
A data frame containing the ROI information.
}
\description{
`r lifecycle::badge("experimental")`

  This function converts detections from template matching or external sources
  into ROIs (Regions of Interest) in the 'monitoraSom' format.
}
