% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_detecs.R
\docType{data}
\name{df_detecs}
\alias{df_detecs}
\title{Detections}
\format{
## `df_detecs` A data frame with 4,363 rows and 21 columns containing
detection results from soundscape analysis, including soundscape and template
file information, template parameters, detection timing and parameters, and
detection peak metrics.
}
\usage{
data(df_detecs)
}
\description{
A data frame containing detections to run the examples of the `monitoraSom`
package. The detections were extracted from the soundscape recordings with
the `run_matching()` function with the default parameters.
}
\keyword{datasets}
