% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_detecs.R
\docType{data}
\name{df_detecs}
\alias{df_detecs}
\title{Detections}
\format{
## `df_detecs` Add description here
}
\usage{
df_detecs
}
\description{
A data frame containing detections to run the examples of the `monitoraSom`
package. The detections were extracted from the soundscape recordings with
the `match_n()` function with the default parameters.
}
\keyword{datasets}
