% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_detecs_val_tovlp.R
\docType{data}
\name{df_detecs_val_tovlp}
\alias{df_detecs_val_tovlp}
\title{Detections with validation labels based on temporal overlap}
\format{
## `df_detecs_val_tovlp` A data frame containing detections with
  validation labels and all the standard detection variables from the
  `run_matching()` function output.
}
\usage{
data(df_detecs_val_tovlp)
}
\description{
A data frame containing detections with validation labels assigned based on
temporal overlap with manually validated detections. The detections were
extracted from soundscape recordings with the `run_matching()` function and
validated by comparing their temporal overlap with manually validated
detections in `df_detecs_val_manual`. Detections with sufficient temporal
overlap with true positives were labeled as true positives (TP), while those
without sufficient overlap were labeled as false positives (FP).
}
\keyword{datasets}
