% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_grid.R
\docType{data}
\name{df_grid}
\alias{df_grid}
\title{Example match grid data}
\format{
## `df_grid` A data frame with 72 rows and 17 columns containing
  soundscape and template file information, including paths, durations,
  sample rates, frequency ranges and other audio parameters used for template
  matching.
}
\usage{
df_grid
}
\description{
A data frame containing example match grid data used to demonstrate the
template matching functionality of the `monitoraSom` package. The match grid
contains the parameters used to extract detections from soundscape
recordings, including the template paths, frequency ranges, and detection
thresholds.
}
\keyword{datasets}
