% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rois.R
\docType{data}
\name{df_rois}
\alias{df_rois}
\title{Concatenated ROI table}
\format{
## `df_rois` add description here
}
\usage{
df_rois
}
\description{
A data frame containing the concatenated ROI tables for the soundscape
recordings after segmentation on the segmentation app, after importing it
back to the R session with the `fetch_roi_tables()` function.
}
\keyword{datasets}
