% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rois.R
\docType{data}
\name{df_rois}
\alias{df_rois}
\title{Concatenated ROI table}
\format{
## `df_rois` A data frame with 79 rows and 19 columns containing ROI
  (region of interest) data from soundscape recordings, including file paths,
  temporal and frequency bounds, labels, and segmentation parameters.
}
\usage{
df_rois
}
\description{
A data frame containing the concatenated ROI tables for the soundscape
recordings after segmentation on the segmentation app, after importing it
back to the R session with the `fetch_roi_tables()` function.
}
\keyword{datasets}
