% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_scores.R
\docType{data}
\name{df_scores}
\alias{df_scores}
\title{Raw results of the template matching}
\format{
## `df_scores` Add description here
}
\usage{
df_scores
}
\description{
A tibble (see the tidyr package for more details) containing the raw results
of the template matching run with the `run_matching()` function with the default
parameters.
}
\keyword{datasets}
