% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_soundscapes.R
\docType{data}
\name{df_soundscapes}
\alias{df_soundscapes}
\title{Metadata of soundscape recordings}
\format{
## `df_soundscapes` Add description here
}
\usage{
df_soundscapes
}
\description{
A data frame containing the metadata of the soundscape recordings to run the
examples of the `monitoraSom` package. The metadata was extracted from the
soundscape recordings with the `fetch_soundscapes_metadata()` function.
}
\keyword{datasets}
