% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_templates.R
\docType{data}
\name{df_templates}
\alias{df_templates}
\title{Metadata of templates}
\format{
## `df_templates` Add description here
}
\usage{
df_templates
}
\description{
A data frame containing the metadata of the templates to run the examples of
the `monitoraSom` package, after importing them back to the R session with
the `fetch_template_metadata()` function.
}
\keyword{datasets}
