% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_validations_i.R
\name{diagnostic_validations_i}
\alias{diagnostic_validations_i}
\title{Diagnostic validation}
\usage{
diagnostic_validations_i(
  val_i,
  diag_method = "Auto",
  pos_prob = 0.95,
  diag_cut = NULL
)
}
\arguments{
\item{val_i}{A data frame containing detections as in the output of
the validation functions.}

\item{diag_method}{A character string indicating the method to select the
peak score cut-off. The default is "Auto". The other option is "Manual".}

\item{pos_prob}{A numeric value between 0 and 1 indicating the probability
of a positive detection. The default is 0.95.}

\item{diag_cut}{A numeric value between 0 and 1 indicating the peak score
cut-off. The default is NULL.}
}
\value{
Validation results.
}
\description{
`r lifecycle::badge("experimental")`

This function takes in a data frame containing detections and validates them
using a binomial regression model. The function returns a data frame with the
validated detections.
}
