% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_validations_i.R
\name{diagnostic_validations_i}
\alias{diagnostic_validations_i}
\title{Compute detection diagnostics of validated detections from one template}
\usage{
diagnostic_validations_i(
  val_i,
  diag_method = "auto",
  pos_prob = 0.95,
  diag_cut = NULL
)
}
\arguments{
\item{val_i}{A tibble containing the validation results of the detections.
This function takes data from a single template only, otherwise it will
return an error.}

\item{diag_method}{A character string indicating the method to use for
diagnostic validations. The two methods available are: "auto" (default) or
"Manual". If "auto" is selected, the function will automatically determine
the cutpoint to be used in the diagnostic validations. If "Manual" is
selected, the user must explicitly pass the cutpoint to be used in the
diagnostic validations to the `diag_cut` argument.}

\item{pos_prob}{A numeric value indicating the probability of a positive
test, if `diag_method` is set to "auto", otherwise it is ignored. It
defaults to 0.95.}

\item{diag_cut}{A numeric value indicating the cutpoint to be used in the
diagnostic, if `diag_method` is set to "manual", otherwise it is ignored.
It defaults to NULL.}
}
\value{
A list containing the diagnostic results
}
\description{
`r lifecycle::badge("experimental")`

  This function performs diagnostic validations on the detections. The
  diagnostic validations are used to determine the optimal cutpoint to be
  used in the detections.
}
