% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_roi_cuts.R
\name{export_roi_cuts}
\alias{export_roi_cuts}
\title{Export ROI cuts}
\usage{
export_roi_cuts(df_rois, roi_cuts_path = "roi_cuts/", overwrite = FALSE)
}
\arguments{
\item{df_rois}{A tibble containing the ROIs to be cut}

\item{roi_cuts_path}{A character string indicating the path to save the cuts.
We recommend to provide the "roi_cuts/" directory, which is the default
path used by the segmentation app.}

\item{overwrite}{If TRUE, existing files will be overwritten}
}
\value{
A message indicating if all cuts were made successfully
}
\description{
`r lifecycle::badge("experimental")`

  This function exports the cuts of the ROIs in the provided ROI table. The
  function will cut the ROIs from the soundscape files and save them in the
  specified path.
}
\examples{
\dontrun{

# Load the necessary packages to run this example
library(monitoraSom)
library(dplyr)

# Load the roi data (output of `fetch_rois()`)
data(df_rois)

# Filter the rois to export as templates. Two layers of filtering are applied:
# 1) the soundscape file paths that contain "recordings" to avoid exporting
# ROIs obtained from the soundscapes recording segmentation.
# 2) the roi comment contains "Substructure C", to select only the ROIs of
# interest
df_rois_processed <- df_rois \%>\%
  filter(grepl("recordings", soundscape_path)) \%>\%
  filter(grepl("Substructure C", roi_comment)) \%>\%
  glimpse()

# Create a directory to store the roi cuts
dir.create("./templates/")

# Export the roi cuts
export_roi_cuts(
  df_rois = df_rois_processed, roi_cuts_path = "./templates/"
)

# Check the roi cuts exported
list.files("./templates/", pattern = "*.wav")

}
}
