% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_roi_cuts_n.R
\name{export_roi_cuts_n}
\alias{export_roi_cuts_n}
\title{Export ROI cuts}
\usage{
export_roi_cuts_n(df_rois, roi_cuts_path = "040_roi_cuts/", overwrite = FALSE)
}
\arguments{
\item{df_rois}{A tibble containing the ROIs to be cut}

\item{roi_cuts_path}{A character string indicating the path to save the cuts.
We recommend to provide the "040_roi_cuts/" directory, which is the default
path used by the segmentation app.}

\item{overwrite}{If TRUE, existing files will be overwritten}
}
\value{
A message indicating if all cuts were made successfully
}
\description{
`r lifecycle::badge("experimental")`

  This function exports the cuts of the ROIs in the provided ROI table. The
  function will cut the ROIs from the soundscape files and save them in the
  specified path.
}
\examples{
\dontrun{

# Load the necessary packages to run this example
library(monitoraSom)
library(dplyr)

# Load the roi data (output of `fetch_rois()`)
data(df_rois)

# Selecting 5 random ROIs to export
df_rois_to_export <- sample_n(df_rois, 5)
glimpse(df_rois_to_export)

# Create a directory to store the roi cuts
dir.create("./temp_roi_cuts/")

# Export the roi cuts
export_roi_cuts_n(
  df_rois = df_rois_to_export, roi_cuts_path = "./temp_roi_cuts/"
)

# Check the roi cuts exported
list.files("./temp_roi_cuts/", pattern = "*.wav")

}
}
