% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_roi_cuts_n.R
\name{export_roi_cuts_n}
\alias{export_roi_cuts_n}
\title{Export ROI cuts}
\usage{
export_roi_cuts_n(rois_n, path)
}
\arguments{
\item{rois_n}{A data frame in the format produced by the segmentation app}

\item{path}{A character string with the path to save the cuts.}
}
\value{
A message indicating if all cuts were made successfully
}
\description{
`r lifecycle::badge("experimental")`

This function takes in a ROI table and export its ROIs as audio files in the
specified path. The ROI table should be in the format produced by the
segmentation app.
}
