% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_roi_cuts_n.R
\name{export_roi_cuts_n}
\alias{export_roi_cuts_n}
\title{Export ROI cuts}
\usage{
export_roi_cuts_n(rois_n, path)
}
\arguments{
\item{rois_n}{A tibble containing the ROIs to be cut}

\item{path}{A character string indicating the path to save the cuts}
}
\value{
A message indicating if all cuts were made successfully
}
\description{
`r lifecycle::badge("experimental")`

  This function exports the cuts of the ROIs in the provided ROI table. The
  function will cut the ROIs from the soundscape files and save them in the
  specified path.
}
