% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_match_grid.R
\name{fetch_match_grid}
\alias{fetch_match_grid}
\title{Creates a grid with all combinations of templates and soundscapes for
template matching}
\usage{
fetch_match_grid(soundscape_data, template_data)
}
\arguments{
\item{soundscape_data}{A data frame containing metadata about the soundscapes
to be matched.}

\item{template_data}{A data frame containing metadata about the templates to
be matched.}
}
\value{
A data frame with a grid of all possible matching combinations
  between the two input data sets. In other words, each row of the output
  data frame represents the match between a template and a soundscape. All
  columns with metadata from soundscapes and templates are kept (see the
  documentation of the functions 'fetch_soundscape_metadata()' and
  'fetch_template_metadata()' for more details).
}
\description{
`r lifecycle::badge("experimental")`

This function takes in two data frames: soundscape_data (the output of the
function 'fetch_soundscape_metadata()') and template_data (the output of the
function 'fetch_template_metadata()'), checks their compatibilities, and
returns a new data frame with a grid of all possible matching combinations
between the two data sets.
}
\examples{
fetch_match_grid(soundscape_data, template_data)
}
