% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_match_grid.R
\name{fetch_match_grid}
\alias{fetch_match_grid}
\title{Fetches a grid of matching templates and soundscapes}
\usage{
fetch_match_grid(soundscape_data, template_data)
}
\arguments{
\item{soundscape_data}{A data frame containing metadata about the soundscapes to be matched.}

\item{template_data}{A data frame containing metadata about the templates to be matched.}
}
\value{
A data frame with a grid of all possible matching combinations between the two input data sets.
}
\description{
This function takes in two data frames: soundscape_data and template_data, and returns a new data frame with a grid of all possible matching combinations between the two data sets.
}
\examples{
fetch_match_grid(soundscape_data, template_data)
}
