% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rois.R
\name{fetch_rois}
\alias{fetch_rois}
\title{Gather ROI tables from a directory}
\usage{
fetch_rois(rois_path = NULL, recursive = FALSE)
}
\arguments{
\item{rois_path}{Path to directory containing ROI tables in CSV format. The
expected variables must be consistent with those exported by the
segmentation app.}

\item{recursive}{Search recursively for roi tables. Defaults to FALSE.}
}
\value{
A data frame with ROI tables
}
\description{
`r lifecycle::badge("experimental")`

  This function gathers ROI tables from a directory and returns a data frame
  with the ROI tables.
}
