% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rois_n.R
\name{fetch_rois_n}
\alias{fetch_rois_n}
\title{Gather ROI tables from a directory}
\usage{
fetch_rois_n(path, recursive = FALSE)
}
\arguments{
\item{path}{Path to directory containing ROI tables in CSV format. The
expected variables must be consistent with those exported by the segmentation
app.}

\item{recursive}{Search recursively for roi tables, default is FALSE}
}
\value{
A data frame with ROI tables
}
\description{
`r lifecycle::badge("experimental")`

  This function gathers ROI tables from a directory and returns a data frame
  with the ROI tables.
}
