% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rois_n.R
\name{fetch_rois_n}
\alias{fetch_rois_n}
\title{ather ROI tables from a directory}
\usage{
fetch_rois_n(path, recursive = FALSE)
}
\arguments{
\item{path}{Path to directory containing ROI tables}

\item{recursive}{Search recursively for roi tables, default is FALSE}
}
\value{
A data frame with ROI tables
}
\description{
ather ROI tables from a directory
}
