% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_score_peaks_i.R
\name{fetch_score_peaks_i}
\alias{fetch_score_peaks_i}
\title{Find score peaks}
\usage{
fetch_score_peaks_i(match_res_i, buffer_size = "template")
}
\arguments{
\item{match_res_i}{A list containing the result of a template matching operation performed with a specific template and soundscape recording. It must contain at least the following elements:
\describe{
  \item{\code{score_vec}}{A list containing the score vector obtained from the matching operation.}
  \item{\code{score_sliding_window}}{The size of the sliding window used in the matching operation.}
  \item{\code{soundscape_path}}{The path to the soundscape recording.}
  \item{\code{soundscape_file}}{The name of the soundscape recording file.}
  \item{\code{template_path}}{The path to the template used in the matching operation.}
  \item{\code{template_file}}{The name of the template file.}
  \item{\code{template_min_freq}}{The minimum frequency of the template.}
  \item{\code{template_max_freq}}{The maximum frequency of the template.}
  \item{\code{template_wl}}{The window length used in the template matching operation.}
  \item{\code{template_ovlp}}{The overlap between windows used in the template matching operation.}
  \item{\code{template_sample_rate}}{The sample rate of the template.}
}}

\item{buffer_size}{An integer indicating the minimum number of points required to define a peak. Defaults to "template", which sets it equal to the sliding window size of the matching operation.}
}
\value{
A data frame with the following columns:
\describe{
  \item{\code{peak_index}}{The index of the peak in the score vector.}
  \item{\code{peak_cor}}{The value of the peak.}
  \item{\code{peak_quant}}{The quantile of the peak value in the score vector.}
  \item{\code{soundscape_path}}{The path to the soundscape recording.}
  \item{\code{soundscape_file}}{The name of the soundscape recording file.}
  \item{\code{template_path}}{The path to the template used in the matching operation.}
  \item{\code{template_file}}{The name of the template file.}
  \item{\code{detection_start}}{The start time of the detection in the soundscape recording.}
  \item{\code{detection_end}}{The end time of the detection in the soundscape recording.}
  \item{\code{min_freq}}{The minimum frequency of the template.}
  \item{\code{max_freq}}{The maximum frequency of the template.}
  \item{\code{detec_wl}}{The window length used in the detection process.}
  \item{\code{detec_ovlp}}{The overlap between windows used in the detection process.}
  \item{\code{detec_sample_rate}}{The sample rate of the soundscape recording.}
  \item{\code{detec_buffer}}{The minimum number of points required to define a peak.}
}
}
\description{
This function detects peaks in the score vector resulting from a template matching operation.
The peaks represent potential detections of a template in a soundscape recording.
}
\examples{
fetch_score_peaks_i(match_res_i, 10)
}
