% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_score_peaks_i.R
\name{fetch_score_peaks_i}
\alias{fetch_score_peaks_i}
\title{Find and filter detections from one score vector}
\usage{
fetch_score_peaks_i(
  df_scores_i,
  buffer_size = "template",
  min_score = NULL,
  min_quant = NULL,
  top_n = NULL
)
}
\arguments{
\item{df_scores_i}{One row of the output of the function 'run_matching()' or the
output of 'run_matching_i()', which contain score vector the result of a template
matching operation performed with a specific template and soundscape
recording.}

\item{buffer_size}{A numeric value specifying the number of frames of the
buffer within which overlap between detections is avoided. Defaults to
"template", which means that the buffer size equals the number of frames
present in the template spectrogram. The buffer exclusion is oriented by
score quantiles, so that the highest scoring detections are always kept.
Setting the buffer size to 0 disables the exclusion buffer.}

\item{min_score}{A numeric value between 0 and 0.99 indicating the minimum
score of the detections that will be kept. Defaults to NULL, which returns
all available detections.}

\item{min_quant}{A numeric value between 0 and 1 indicating the minimum score
quantile of the kept detections. Defaults to NULL, which returns all
available detections.}

\item{top_n}{An integer indicating the maximum number of peaks to be
returned, selected according to the highest scores available. Defaults to
NULL, which return all available detections. It should be noted that
because the peak quantiles are callculated within each score vector, the
top_n parameter is applied to each score vector separately, and not to the
whole matching grid.}
}
\value{
A data frame in which each row is a detection and has the follwing
  attributes:
}
\description{
`r lifecycle::badge("experimental")`

  This function detects peaks in the score vector resulting from one
  iteration of template matching, i.e. the output of 'run_matching_i()' or one row
  of the output of 'run_matching()' . The peaks represent potential detections of
  a template in a soundscape recording.
}
