% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_score_peaks_n.R
\name{fetch_score_peaks_n}
\alias{fetch_score_peaks_n}
\title{Extract score peaks from a Tibble of audio scores}
\usage{
fetch_score_peaks_n(
  tib_match,
  tib_match_path = NULL,
  buffer_size = "template",
  save_res = NULL
)
}
\arguments{
\item{tib_match}{Tibble. Tibble containing audio scores obtained with \code{\link{match_template_n}}.}

\item{tib_match_path}{Character. Path to a folder containing audio scores as .wav.rds files.}

\item{buffer_size}{Integer. Size of the buffer around the detected peak (in number of points).}

\item{save_res}{Character. Path to save the result as an .rds file.}
}
\value{
A Tibble containing the detections of all audio scores.
}
\description{
This function extracts score peaks from a Tibble of audio scores obtained with \code{\link{match_template_n}}.
}
\examples{
\dontrun{
# Load example audio scores
tib_match <- readRDS(system.file("extdata", "example_audio_scores.rds", package = "my_package"))

# Extract detections from scores
tib_detecs <- fetch_score_peaks_n(tib_match, buffer_size = 25, save_res = "detections.rds")
}

}
\seealso{
\code{\link{match_template_n}}, \code{\link{fetch_score_peaks_i}}
}
