% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_soundscape_metadata.R
\name{fetch_soundscape_metadata}
\alias{fetch_soundscape_metadata}
\title{Fetch Soundscape Metadata}
\usage{
fetch_soundscape_metadata(path, ncores = 1)
}
\arguments{
\item{path}{A character string specifying the directory path where the WAV files are located.}

\item{ncores}{An integer specifying the number of CPU cores to use for parallel processing.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{soundscape_path}}{A character string specifying the full path to the WAV file.}
\item{\code{soundscape_file}}{A character string specifying the name of the WAV file.}
\item{\code{soundscape_duration}}{A numeric value specifying the duration of the WAV file in seconds.}
\item{\code{soundscape_sample_rate}}{An integer specifying the sample rate of the WAV file in Hz.}
\item{\code{soundscape_codec}}{A character string specifying the audio codec used in the WAV file.}
\item{\code{soundscape_layout}}{A character string specifying the audio channel layout of the WAV file# .}
}
}
\description{
Extracts metadata from all WAV files found in a directory and its subdirectories.
}
\examples{
fetch_soundscape_metadata(path = "/path/to/soundscapes", ncores = 4)

}
