% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_soundscape_metadata.R
\name{fetch_soundscape_metadata}
\alias{fetch_soundscape_metadata}
\title{Extract Soundscape Metadata}
\usage{
fetch_soundscape_metadata(path, recursive = TRUE, ncores = 1)
}
\arguments{
\item{path}{A character string specifying the directory path where all WAV
files are located.}

\item{recursive}{A logical value indicating whether the search for WAV files
should be recursive or not. Default is TRUE.}

\item{ncores}{An integer indicating the number of cores to be used for
parallelization. Default is 1.}
}
\value{
A data frame with the following columns:
}
\description{
`r lifecycle::badge("experimental")`

Extracts metadata from all WAV files found in a directory and its
subdirectories.
}
