% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_template_metadata.R
\name{fetch_template_metadata}
\alias{fetch_template_metadata}
\title{Extract template metadata}
\usage{
fetch_template_metadata(path, method = c("standalone"))
}
\arguments{
\item{path}{The directory path containing the template waveform files.}

\item{method}{The method to use for metadata extraction. Options include "standalone" to extract metadata from standalone waveform files, or "roi_table" to extract metadata from a table of regions of interest (ROIs).}
}
\value{
A tibble containing the following metadata for each template file:
\itemize{
\item \code{template_path}: The full path to the template file.
\item \code{template_file}: The filename of the template file.
\item \code{template_label}: A label for the template file, extracted from the filename.
\item \code{template_start}: The start time of the template file, set to 0.
\item \code{template_end}: The end time of the template file, calculated from the duration of the file.
\item \code{template_sample_rate}: The sample rate of the template file.
\item \code{template_min_freq}: The minimum frequency of the template file, extracted from the filename.
\item \code{template_max_freq}: The maximum frequency of the template file, extracted from the filename.
\item \code{template_wl}: The waveform length (WL) of the template file, extracted from the filename.
\item \code{template_ovlp}: The waveform overlap (OVLP) of the template file, extracted from the filename.
}
}
\description{
This function extracts metadata from a directory of template waveform files, which can be used for soundscape analysis.
}
\examples{
# Extract metadata from standalone waveform files
fetch_template_metadata("/path/to/templates", method = "standalone")

# Extract metadata from a table of ROIs
fetch_template_metadata("/path/to/rois.csv", method = "roi_table")
}
