% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_template_metadata.R
\name{fetch_template_metadata}
\alias{fetch_template_metadata}
\title{Extract template metadata}
\usage{
fetch_template_metadata(path, recursive = TRUE, method = c("standalone"))
}
\arguments{
\item{path}{The directory path containing the template waveform files (for method = "standalone") or CSV files (for method = "roi_table").}

\item{recursive}{A logical value indicating whether the search for template files should be recursive or not. Default is TRUE.}

\item{method}{The method of metadata extraction. The available Options include "standalone" and "roi_table". See details for more information on the format of the input files for each method.}
}
\value{
A tibble containing the following metadata for each template file:
}
\description{
`r lifecycle::badge("experimental")`

This function extracts metadata from a directory of template waveform files, which can be used for soundscape analysis.
}
