% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_template_metadata.R
\name{fetch_template_metadata}
\alias{fetch_template_metadata}
\title{Extract template metadata}
\usage{
fetch_template_metadata(templates_path = NULL, recursive = FALSE)
}
\arguments{
\item{templates_path}{The directory path containing the template wave files.
If no path is provided, the function will look for the templates in the
"roi_cuts/" directory.}

\item{recursive}{A logical value indicating whether the search for template
files should be recursive or not. Defaults to FALSE.}
}
\value{
A tibble containing the following metadata for each template file:
}
\description{
`r lifecycle::badge("experimental")`

  This function extracts metadata from a directory of template wave files,
  which can be used for soundscape analysis.
}
\examples{
\dontrun{

# Load the necessary packages to run this example
library(monitoraSom)
library(dplyr)

# Load the roi data (output of `fetch_rois()`)
data(df_rois)

# Filter the rois to export as templates. Two layers of filtering are applied:
# 1) the soundscape file paths that contain "recordings" to avoid exporting
# ROIs obtained from the soundscapes recording segmentation.
# 2) the roi comment contains "Substructure C", to select only the ROIs of
# interest
df_rois_processed <- df_rois \%>\%
  filter(grepl("recordings", soundscape_path)) \%>\%
  filter(grepl("Substructure C", roi_comment)) \%>\%
  glimpse()

# Create a directory to store the roi cuts
dir.create("./templates/")

# Export the roi cuts
export_roi_cuts_n(
  df_rois = df_rois_processed, roi_cuts_path = "./templates/"
)

# Import the templates metadata
df_templates <- fetch_template_metadata(templates_path = "./templates/")
glimpse(df_templates)

}
}
