% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_template_metadata.R
\name{fetch_template_metadata}
\alias{fetch_template_metadata}
\title{Extract template metadata}
\usage{
fetch_template_metadata(templates_path = NULL, recursive = FALSE)
}
\arguments{
\item{templates_path}{The directory path containing the template wave files.
If no path is provided, the function will look for the templates in the
"040_roi_cuts/" directory.}

\item{recursive}{A logical value indicating whether the search for template
files should be recursive or not. Defaults to FALSE.}
}
\value{
A tibble containing the following metadata for each template file:
}
\description{
`r lifecycle::badge("experimental")`

  This function extracts metadata from a directory of template wave files,
  which can be used for soundscape analysis.
}
\examples{
\dontrun{

# Load the necessary packages to run this example
library(monitoraSom)
library(dplyr)
library(tuneR)

# Load the templates to populate the example data
data(ls_templates)

# Create a directory and export the templates
templates_path <- "./040_roi_cuts"
dir.create(templates_path)
invisible(lapply(1:length(ls_templates), function(i) {
  writeWave(
    ls_templates[[i]], file.path(templates_path, names(ls_templates)[i])
  )
}))

# Import the templates metadata
df_templates <- fetch_template_metadata(templates_path = templates_path)
glimpse(df_templates)

}
}
