% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_external_detections.R
\name{import_external_detections}
\alias{import_external_detections}
\title{Import external detections}
\usage{
import_external_detections(input_path, names_vec = NA, output_path = NA)
}
\arguments{
\item{input_path}{A character string indicating the path to the CSV file}

\item{names_vec}{A named character vector indicating the names of the
variables and the corresponding names in the provided table.}

\item{output_path}{A character string indicating the path to save the output}
}
\value{
A tibble containing the detections in the 'monitoraSom' format
}
\description{
`r lifecycle::badge("experimental")`

  This function imports detections from an external CSV file and converts
  them to the 'monitoraSom' format. The function will rename the variables to
  match the 'monitoraSom' format and will add the necessary variables to
  complete the format.
}
