% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_segmentation_app.R
\name{launch_segmentation_app}
\alias{launch_segmentation_app}
\title{Function to launch the segmentation app}
\usage{
launch_segmentation_app(
  project_path = NULL,
  preset_path = NULL,
  user = NULL,
  soundscapes_path = NULL,
  roi_tables_path = NULL,
  cuts_path = NULL,
  labels_file = NULL,
  sp_list = "CBRO-2021 (Birds - Brazil)",
  label_angle = 90,
  show_label = TRUE,
  dyn_range = c(0, 50),
  wl = 1024,
  ovlp = 0,
  color_scale = "inferno",
  wav_player_type = "HTML player",
  wav_player_path = "play",
  visible_bp = FALSE,
  play_norm = FALSE,
  session_notes = NULL,
  zoom_freq = c(0, 180),
  nav_autosave = TRUE,
  pitch_shift = 1
)
}
\arguments{
\item{project_path}{Path to the project folder, where the segmentation
project will be stored. It is necessary to provide this path if the
segmentation project is being created for the first time or if the user
wants to use relative paths for other input fields.}

\item{preset_path}{Path to temporary files and other acessory files used by the app.}

\item{user}{Identification of the user of the segmentation app.}

\item{soundscapes_path}{Path to soundscapes files.}

\item{roi_tables_path}{Path to the folder from which the ROI tables will be
imported and to which they will be exported.}

\item{cuts_path}{Path to the folder from which the ROI audio cuts will be
imported and to which they will be exported.}

\item{labels_file}{Path to the file containing the list of labels to be used
in the segmentation. This file must be a '.xlsx' spreadsheet in which
available lists are stored as columns, identified in the app menu by their
titles.}

\item{sp_list}{A character string with the available labels to be used in the
cuirrent session. The default is "CBRO-2021 (Brazil)".}

\item{label_angle}{Angle between 0 and 180 to draw the ROI labels in the
spectrogram plot.}

\item{show_label}{If TRUE, ROI labels will be displayed alongside ROI
selections in the sepctrogram.}

\item{dyn_range}{A numeric vector of length 2 between -100 and 0, specifying
the minimum and maximum relative amplitudes to be displayed in the
spectrogram}

\item{wl}{An integer specifying the length of the FFT window used to
calculate the spectrogram.}

\item{ovlp}{A numeric value between 0 and 90 specifying the percentage
overlap of windows for computing the spectrogram.}

\item{color_scale}{A character string specifying the color scale to be used
in the spectrogram. Possible values are "viridis", "magma", "inferno",
"cividis", "greyscale 1", or "greyscale 2".}

\item{wav_player_type}{The type of wav player. "R session" for R
session-based player, "HTML player" for an embedded HTML player in the
interface, and "External player" for playing in an external player. The
last options becomes available only when a valid path to the player
executable is provided in the 'wav_player_path' argument.}

\item{wav_player_path}{Path to the wav player executable (only for
wav_player_type = "External player")}

\item{visible_bp}{If TRUE, the visible frequency band will be used as a
bandpass filter for the audio playback.}

\item{session_notes}{A single character string with related to the current
segmentation session.}

\item{zoom_freq}{Vector with two numeric values between 0 and the Nyquist
Frequency of the soundscape recording, indicating the frequency values in
kHz of the frequency band to be displayed in the spectrogram.}

\item{nav_autosave}{If TRUE, navigating between soundscapes will
automatically save the ROI table of the active soundscape.}

\item{pitch_shift}{A numeric value indicating the pitch shift to be applied
to the soundscape audio. The default is 1, which means no pitch shift.}
}
\value{
Todo
}
\description{
`r lifecycle::badge("experimental")`

  This function launches the segmentation app, which is a Shiny app for
  segemntation of WAV recorcings of soundscapes into tables containing
  regions of interest (ROIs) and audio cuts of the ROIs. The app settings can
  be imported from presets or set manually.
}
\examples{
\dontrun{
library(monitoraSom)
library(usethis)

# set the path to the diorectory where soundscapes are located (it is not recursive)
soundscapes_path <- "path/to/soundscapes"
# set the path to the project folder
project_path <- "path/to/project"

# in case the current working directory is not an active project, set it with the following command. it may be necessary to restart the R session after that.
create_project(path = "MyProject", open = TRUE, rstudio = TRUE)

# launch the segmentation app
launch_segmentation_app(
  project_path = ".",
  user = "Identify the user here",
  soundscapes_path = soundscapes_path
)
}
}
