% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_segmentation_app.R
\name{launch_segmentation_app}
\alias{launch_segmentation_app}
\title{Function to launch the segmentation app}
\usage{
launch_segmentation_app(
  project_path = NULL,
  preset_path = NULL,
  user = NULL,
  soundscapes_path = NULL,
  roi_tables_path = NULL,
  cuts_path = NULL,
  labels_file = NULL,
  sp_list = "CBRO-2021 (Birds - Brazil)",
  label_angle = 90,
  show_label = TRUE,
  time_guide_interval = 3,
  freq_guide_interval = 2,
  dyn_range = c(-60, 0),
  dyn_range_bar = c(-144, 0),
  wl = 1024,
  ovlp = 0,
  color_scale = "inferno",
  wav_player_type = "HTML player",
  wav_player_path = "play",
  visible_bp = FALSE,
  play_norm = FALSE,
  session_notes = NULL,
  zoom_freq = c(0, 180),
  nav_autosave = TRUE,
  pitch_shift = 1,
  confirm_paths = TRUE
)
}
\arguments{
\item{project_path}{Path to the project folder, where the segmentation
project will be stored. It is necessary to provide this path if the
segmentation project is being created for the first time or if the user
wants to use relative paths for other input fields.}

\item{preset_path}{Path to temporary files and other acessory files used by
the app.}

\item{user}{Identification of the user of the segmentation app.}

\item{soundscapes_path}{Path to soundscapes files.}

\item{roi_tables_path}{Path to the folder from which the ROI tables will be
imported and to which they will be exported.}

\item{cuts_path}{Path to the folder from which the ROI audio cuts will be
imported and to which they will be exported.}

\item{labels_file}{Path to the file containing the list of labels to be used
in the segmentation. This file must be a '.xlsx' spreadsheet in which
available lists are stored as columns, identified in the app menu by their
titles.}

\item{sp_list}{A character string with the available labels to be used in the
current session. The default is "CBRO-2021 (Birds - Brazil)".}

\item{label_angle}{Angle between 0 and 180 to draw the ROI labels in the
spectrogram plot.}

\item{show_label}{If TRUE, ROI labels will be displayed alongside ROI
selections in the sepctrogram.}

\item{time_guide_interval}{A numeric value indicating the interval in seconds
between time guides in the spectrogram.}

\item{freq_guide_interval}{A numeric value indicating the interval in kHz
between frequency guides in the spectrogram.}

\item{dyn_range}{A numeric vector of length 2 specifying the minimum and
maximum relative amplitudes (in dBFS) to be displayed on the spectrogram.
By default c(-100, 0).}

\item{dyn_range_bar}{A numeric vector of length 2 specifying the limits to be
diplayed on the the bar that controls the dinamic range on the app. By
default c(-200, 10).}

\item{wl}{An integer specifying the length of the FFT window used to
calculate the spectrogram.}

\item{ovlp}{A numeric value between 0 and 90 specifying the percentage
overlap of windows for computing the spectrogram.}

\item{color_scale}{A character string specifying the color scale to be used
in the spectrogram. Possible values are "viridis", "magma", "inferno",
"cividis", "greyscale 1", or "greyscale 2".}

\item{wav_player_type}{The type of wav player. "HTML player" for an embedded
HTML player in the interface, "R session" (experimental) for R
session-based player, and "External player" (experimental) for playing in
an external player. The last options becomes available only when a valid
path to the player executable is provided in the 'wav_player_path'
argument.}

\item{wav_player_path}{Path to the wav player executable (only for
wav_player_type = "External player")}

\item{visible_bp}{If TRUE, the visible frequency band will be used as a
bandpass filter for the audio playback.}

\item{play_norm}{If TRUE, the audio playback will be normalized.}

\item{session_notes}{A single character string with related to the current
segmentation session.}

\item{zoom_freq}{Vector with two numeric values between 0 and the Nyquist
Frequency of the soundscape recording, indicating the frequency values in
kHz of the frequency band to be displayed in the spectrogram.}

\item{nav_autosave}{If TRUE, navigating between soundscapes will
automatically save the ROI table of the active soundscape.}

\item{pitch_shift}{A numeric value indicating the pitch shift to be applied
to the soundscape audio. The default is 1, which means no pitch shift.}

\item{confirm_paths}{If TRUE, the user will be asked to confirm the paths
setup at the app startup.}
}
\value{
A pop up with the rendered app.
}
\description{
`r lifecycle::badge("experimental")`

  This function launches the segmentation app, which is a Shiny app for
  segemntation of WAV recorcings of soundscapes into tables containing
  regions of interest (ROIs) and audio cuts of the ROIs. The app settings can
  be imported from presets or set manually.
}
\examples{
\dontrun{

# Load the necessary packages to run this example
library(monitoraSom)

# Essential setup
launch_segmentation_app(
  project_path = ".", user = "User",
  soundscapes_path = "./soundscapes", roi_tables_path = "./roi_tables/",
  cuts_path = "./roi_cuts/"
)

# Check the segmentation files produced by the segmentation app
list.files("./roi_cuts/", pattern = "*.csv")

# Check the roi cuts exported to be used as templates
list.files("./roi_cuts/", pattern = "*.wav")

# Detailed setup
launch_segmentation_app(
  project_path = ".", user = "User",
  soundscapes_path = "./soundscapes", roi_tables_path = "./roi_tables/",
  cuts_path = "./roi_cuts/",
  dyn_range = c(-84, -48), wl = 1024, ovlp = 80, zoom_freq = c(0, 10),
  color_scale = "greyscale 1",
)

}
}
