% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_validation_app.R
\name{launch_validation_app}
\alias{launch_validation_app}
\title{Launch validation app}
\usage{
launch_validation_app(
  project_path = NULL,
  preset_path = NULL,
  validation_user,
  templates_path,
  soundscapes_path,
  input_path,
  output_path = NULL,
  spec_path = NULL,
  wav_cuts_path = NULL,
  diag_tab_path = NULL,
  wav_player_path = "play",
  wav_player_type = "HTML player",
  val_subset = c("NA", "TP", "FP", "UN"),
  min_score = 0,
  time_pads = 1,
  ovlp = 0,
  wl = 2048,
  dyn_range = c(0, 50),
  color_scale = "inferno",
  zoom_freq = c(0, 10),
  nav_shuffle = FALSE,
  seed = 123,
  auto_next = FALSE,
  nav_autosave = FALSE,
  overwrite = FALSE,
  pitch_shift = 1,
  visible_bp = FALSE,
  play_norm = FALSE
)
}
\arguments{
\item{preset_path}{Path from which presets can be imported and to which new
presets can be exported.}

\item{validation_user}{User name.}

\item{templates_path}{Path to the template wave files.}

\item{soundscapes_path}{Path to the soundscape wave files.}

\item{input_path}{Path to the input file containing detections.}

\item{output_path}{Path to the output file.}

\item{spec_path}{Path to the folder containing the spectrogram images.}

\item{wav_cuts_path}{Path to the folder containing the cut wave files.}

\item{diag_tab_path}{Path to the folder containing the diagnostic tables.}

\item{wav_player_path}{Path to the wav player executable (only for system
player).}

\item{wav_player_type}{The type of wav player. "R session" for R
session-based player, "system" for system player.}

\item{val_subset}{Subset of detections to be validated.}

\item{min_score}{Minimum score of the detections to be validated.}

\item{time_pads}{Time pads to be added to the start and end of the cut wave
files.}

\item{ovlp}{Overlap between consecutive cuts.}

\item{wl}{Window length for the spectrogram.}

\item{dyn_range}{Dynamic range for the spectrogram.}

\item{color_scale}{Color scale for the spectrogram.}

\item{zoom_freq}{Frequency range to zoom in the spectrogram.}

\item{nav_shuffle}{If TRUE, the files will be shuffled before navigation.}

\item{seed}{Seed for the random shuffling.}

\item{auto_next}{If TRUE, the next file is automatically displayed when the
user validates a cut.}

\item{nav_autosave}{If TRUE, the current validation is saved when the user
navigates to another file.}

\item{overwrite}{If TRUE, the output file is overwritten.}

\item{session_notes}{Notes related to the validation session.}
}
\value{
todo
}
\description{
`r lifecycle::badge("experimental")`

  This function launches the validation app, which is a Shiny app to validate
  for segemntation of WAV recorcings of soundscapes into tables containing
  regions of interest (ROIs) and audio cuts of the ROIs. The app settings can
  be imported from presets or set manually.
}
\examples{
\dontrun{
library(monitoraSom)
library(usethis)

# set the path to the diorectory where soundscapes are located (it is not recursive)
soundscapes_path <- "path/to/soundscapes"
# set the path to the project folder
project_path <- "path/to/project"

# in case the current working directory is not an active project, set it with the following command. it may be necessary to restart the R session after that.
create_project(path = paste0(project_path, "/MyProject", open = TRUE, rstudio = TRUE)

# launch the segmentation app
launch_segmentation_app(
  project_path = project_path,
  user = "Identify the user here",
  soundscapes_path = soundscapes_path
)
}
}
