% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_validation_app.R
\name{launch_validation_app}
\alias{launch_validation_app}
\title{Launch validation app}
\usage{
launch_validation_app(
  project_path = ".",
  preset_path = NULL,
  validation_user,
  templates_path = NULL,
  soundscapes_path = NULL,
  input_path = NULL,
  output_path = NULL,
  detec_spec_path = NULL,
  detec_cuts_path = NULL,
  wav_player_path = "play",
  wav_player_type = "HTML player",
  val_subset = c("NV", "TP", "FP", "UN"),
  time_pads = 1,
  ovlp = 0,
  wl = 2048,
  dyn_range_bar = c(-144, 0),
  dyn_range_templ = c(-84, 0),
  dyn_range_detec = c(-84, 0),
  color_scale = "inferno",
  zoom_freq = c(0, 23),
  time_guide_interval = 1,
  freq_guide_interval = 1,
  nav_shuffle = FALSE,
  subset_seed = 123,
  auto_next = TRUE,
  nav_autosave = TRUE,
  overwrite = FALSE,
  pitch_shift = 1,
  visible_bp = FALSE,
  play_norm = FALSE
)
}
\arguments{
\item{project_path}{Path to the project folder. Defaults to ".".}

\item{preset_path}{Path from which presets can be imported and to which new
presets can be exported.}

\item{validation_user}{User name.}

\item{templates_path}{Path to the template wave files. Defaults to
"roi_cuts/". This directory must contain at least one `.wav` file for the
app to work.}

\item{soundscapes_path}{Path to the soundscape wave files. Defaults to
"soundscapes/". This directory must contain at least one `.wav` file for
the app to work.}

\item{input_path}{Path to the input file containing detections.}

\item{output_path}{Path to the output file. Defaults to NULL.}

\item{detec_spec_path}{Path to the folder containing the spectrogram images.
Defaults to "detection_spectrograms/".}

\item{detec_cuts_path}{Path to the folder containing the cut wave files.
Defaults to "detection_cuts/".}

\item{wav_player_path}{Path to the wav player executable (only for system
player). Defaults to "play".}

\item{wav_player_type}{The type of wav player. "R session" for R
session-based player, "system" for system player. Defaults to "HTML
player".}

\item{val_subset}{Subset of detections to be validated.}

\item{time_pads}{Time pads to be added to the start and end of the cut wave
files.}

\item{ovlp}{Overlap between consecutive cuts.}

\item{wl}{Window length for the spectrogram.}

\item{dyn_range_bar}{Adjustment of the maximum range of the dynamic range
slider.}

\item{dyn_range_templ}{Dynamic range for the template spectrogram.}

\item{dyn_range_detec}{Dynamic range for the detection spectrogram.}

\item{color_scale}{Color scale for the spectrogram.}

\item{zoom_freq}{Frequency range to zoom in the spectrogram.}

\item{time_guide_interval}{A numeric value indicating the interval in seconds
between time guides in the spectrogram.}

\item{freq_guide_interval}{A numeric value indicating the interval in kHz
between frequency guides in the spectrogram.}

\item{nav_shuffle}{If TRUE, the files will be shuffled before navigation.}

\item{subset_seed}{Seed for the random shuffling.}

\item{auto_next}{If TRUE, the next file is automatically displayed when the
user validates a cut.}

\item{nav_autosave}{If TRUE, the current validation is saved when the user
navigates to another file.}

\item{overwrite}{If TRUE, the output file is overwritten.}

\item{pitch_shift}{Pitch shift for the audio cuts.}

\item{visible_bp}{If TRUE, the bandpass filter is visible in the spectrogram.}

\item{play_norm}{If TRUE, the played audio is normalized.}
}
\value{
todo
}
\description{
`r lifecycle::badge("experimental")`

  This function launches the validation app, which is a Shiny app to validate
  for segemntation of WAV recorcings of soundscapes into tables containing
  regions of interest (ROIs) and audio cuts of the ROIs.
}
\examples{
\dontrun{
library(monitoraSom)
library(usethis)

# set the path to the diorectory where soundscapes are located (it is not
# recursive)
soundscapes_path <- "path/to/soundscapes"
# set the path to the project folder
project_path <- "path/to/project"

# in case the current working directory is not an active project, set it with
# the following command. it may be necessary to restart the R session after that.
create_project(path = paste0(project_path, "/MyProject", open = TRUE, rstudio = TRUE)

# launch the segmentation app
launch_segmentation_app(
  project_path = project_path,
  user = "Identify the user here",
  soundscapes_path = soundscapes_path
)
}
}
