% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_recordings.R
\docType{data}
\name{ls_recordings}
\alias{ls_recordings}
\title{List of templates}
\format{
## `ls_recordings` A list of `tuneR::Wave` objects containing focal
  recordings. Names of the list elements are the names of the `.wav` files,
  and correspond to the name code for recordings in the `monitoraSom`
  workflow.
}
\usage{
ls_recordings
}
\description{
A list of two focal recordings of Basileuterus culicivorus from which the
templates of the `monitoraSom` package examples were derived.
}
\keyword{datasets}
