% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_roi_tables.R
\docType{data}
\name{ls_roi_tables}
\alias{ls_roi_tables}
\title{List of ROI tables}
\format{
## `ls_roi_tables` A list of twelve `data.frame` objects containing
  the ROI tables for the soundscape recordings after segmentation on the
  segmentation app. Names of the list elements are the names of the `.csv`
  files. The list includes both the segmentation of the focal recordings to
  make the templates as well as the segmentation of the soundscape recordings
  to validate the detections.
}
\usage{
data(ls_roi_tables)
}
\description{
A list of ROI tables to run the examples of the `monitoraSom` package.
}
\keyword{datasets}
