% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_soundscapes.R
\docType{data}
\name{ls_soundscapes}
\alias{ls_soundscapes}
\title{List of soundscape recordings}
\format{
## `ls_soundscapes` A list of twelve `tuneR::Wave` objects containing
  (mostly) target signals from Basileuterus culicivorus to run the examples
  of the `monitoraSom` package. All recordings are from different points of
  the same location, but from different days and times. Names of the list
  elements are the names of the soundscape recordings.
}
\usage{
ls_soundscapes
}
\description{
A list of soundscape recordings to run the examples of the `monitoraSom`
package.
}
\keyword{datasets}
