% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_templates.R
\docType{data}
\name{ls_templates}
\alias{ls_templates}
\title{List of templates}
\format{
## `ls_templates` A list of six `tuneR::Wave` objects containing
  templates to run the examples of the `monitoraSom` package. All templates
  contain signals correspoding to the last two notes of Basileuterus
  culicivorus' song. Names of the list elements are the names of the `.wav`
  files, and correspondo to the name code for templates in the `monitoraSom`
  workflow.
}
\usage{
data(ls_templates)
}
\description{
A list of templates to run the examples of the `monitoraSom` package.
}
\keyword{datasets}
