% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_n.R
\name{match_n}
\alias{match_n}
\title{Match templates in a grid of images}
\usage{
match_n(df_grid, score_method = "cor", ncores = 1, save_res = FALSE)
}
\arguments{
\item{df_grid}{A data frame containing the paths of the images to be analyzed.}

\item{score_method}{A character string indicating the method to be used to calculate the matching score. Default is "cor".}

\item{ncores}{An integer indicating the number of cores to be used for parallelization. Default is 1.}

\item{save_res}{A character string indicating the path to save the results as an RDS file. Default is FALSE.}
}
\value{
A data frame containing the matching scores for each template in the grid.
}
\description{
This function matches templates in a grid of images using the `match_i` function. O chatGPT viajou na maionese, não esquecer de corrigir isso aqui!!!
}
\examples{
# Load example data
data("df_grid")

# Match templates
res <- match_n(df_grid, score_method = "ssim", ncores = 2, save_res = "res.rds")
}
