% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_n.R
\name{match_n}
\alias{match_n}
\title{Batch template matching}
\usage{
match_n(
  df_grid,
  score_method = "cor",
  save_res = FALSE,
  par_strat = "future",
  ncores = 1,
  backend_type = "async",
  cluster_type = "psock"
)
}
\arguments{
\item{df_grid}{The output of the function 'fetch_match_grid()}

\item{score_method}{A character string indicating the method to use for
matching. The two methods available are: "cor" (Pearson correlation
coefficient) or "dtw" (dynamic time warping). Defaults to "cor".}

\item{save_res}{A character string indicating the path to save the results in
the format of an RDS file. Default is FALSE.}

\item{par_strat}{A character string indicating the parallelization strategy
to be used. The available options are: "foreach" (default), "future" and
"pbapply". The 'future' and 'pbapply' strategies do not work on Windows,
but are more efficient in linux (especially when running and R session
outside of Rstudio). The 'foreach' strategy works on all platforms, but is
less efficient than the other two. See the documentation of the 'future'
and 'pbapply' packages for more details.}

\item{ncores}{An integer indicating the number of cores to be used for
parallelization. Default is 1.}
}
\value{
A tibble containing input data frame with an additional column
  "score_vec", which is a list of dataframes with the columns "time_vec" (the
  time value of each spectrogram frame) and "score_vec" (the matching score
  obtained when the template and the soundscape spectrogram of samew
  dimensions are alligned at that frame) for each match. The length of the
  "score_vec" is equal to the number of frames of the soundscape spectrogram
  minus the number of frames of the template spectrogram (i.e. the number of
  possible allignments between the two spectrograms. The score is not
  available for the first and last frames of the soundscape spectrogram
  because score cannot be calculated between spectrograms of different
  dimensions. To produce a score vector with the same number of frames of the
  soundscape spectrogram, pads with length quals half the number of frames
  from the template are added to the beginning and end of the
}
\description{
`r lifecycle::badge("experimental")`

This function is wrapper of 'match_i()' to perform a batch computation of the
matching vectors of the template and soundscape matches defined in the
metadata grid obtained as output of the function 'fetch_match_grid()'. The
available algorithms to compare spectrograms and compute matching scores are
the Pearson correlation coefficient ("cor") or dynamic time warping ("dtw").
}
\examples{
# Load example data
data("df_grid")

# Match templates
res <- match_n(df_grid, score_method = "cor", ncores = 2, save_res = "res.rds")

}
