% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_matching_i.R
\name{run_matching_i}
\alias{run_matching_i}
\title{Run one iteration of the matching algorithm to obtain the matching score
vector between one template and one soundscape}
\usage{
run_matching_i(
  df_grid_i,
  score_method = "cor",
  output = "detections",
  buffer_size = "template",
  min_score = NULL,
  min_quant = NULL,
  top_n = NULL
)
}
\arguments{
\item{df_grid_i}{One row of the output of the function 'fetch_match_grid()}

\item{score_method}{A character string indicating the method to use for
matching. The two methods available are: "cor" (Pearson correlation
coefficient) or "dtw" (dynamic time warping). Defaults to "cor".}

\item{output}{A character string indicating the output of the function. The
two options are: "detections" (default) or "scores".}

\item{buffer_size}{A character string indicating the size of the buffer.}

\item{min_score}{Minimum score threshold (numeric between 0 and 1).}

\item{min_quant}{Minimum quantile threshold (numeric between 0 and 1).}

\item{top_n}{Number of top detections to return (positive integer).}
}
\value{
A tibble containing either detection results or raw scores.
}
\description{
`r lifecycle::badge("experimental")`

  This function takes uses the metadata contained in one row of the output of
  the function 'fetch_match_grid()' to calculate the matching score of the
  template spectrogram and a portion of the soundscape spectrogram of same
  dimensions. The available matching algorithms are the Pearson correlation
  coefficient ("cor") or dynamic time warping ("dtw").
}
