% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_workspace.R
\name{set_workspace}
\alias{set_workspace}
\title{Set the workspace for the monitoraSom package}
\usage{
set_workspace(
  project_path = NULL,
  ext_soundscapes_path = FALSE,
  soundscapes_path = "./soundscapes/",
  soundscapes_metadata_path = "./soundscapes_metadata/",
  roi_tables_path = "./roi_tables/",
  roi_cuts_path = "./roi_cuts/",
  detections_path = "./detections/",
  validation_outputs_path = "./validation_outputs/",
  detection_spectrograms_path = "./detection_spectrograms/",
  detection_cuts_path = "./detection_cuts/",
  validation_diagnostics_path = "./validation_diagnostics/",
  app_presets_path = "./app_presets/"
)
}
\arguments{
\item{project_path}{The path to the project directory. If NULL, the function
will try to determine the project path automatically based on the path to
the R script that is being executed.}

\item{ext_soundscapes_path}{Logical indicating whether the soundscapes path
is external to the project directory. If TRUE, the soundscapes directory
will not be created.}

\item{soundscapes_path}{The path to the soundscapes directory.}

\item{soundscapes_metadata_path}{The path to the soundscapes metadata
directory.}

\item{roi_tables_path}{The path to the roi tables directory.}

\item{roi_cuts_path}{The path to the roi cuts directory.}

\item{detections_path}{The path to the detections directory.}

\item{validation_outputs_path}{The path to the validation outputs directory.}

\item{detection_spectrograms_path}{The path to the detection spectrograms
directory.}

\item{detection_cuts_path}{The path to the detection cuts directory.}

\item{validation_diagnostics_path}{The path to the validation diagnostics
directory.}

\item{app_presets_path}{The path to the app presets directory.}
}
\description{
`r lifecycle::badge("experimental")`

This function sets the working directory to the project path and creates
the necessary directories if they do not exist. It also creates the species
labels file if it doesn't exist.
}
