% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_workspace.R
\name{set_workspace}
\alias{set_workspace}
\title{Set the workspace for the monitoraSom package}
\usage{
set_workspace(
  project_path = NULL,
  example_data = FALSE,
  app_presets_path = "./app_presets/",
  soundscapes_path = "./soundscapes/",
  soundscapes_metadata_path = "./soundscapes_metadata/",
  recordings_path = "./recordings/",
  roi_tables_path = "./roi_tables/",
  roi_cuts_path = "./roi_cuts/",
  templates_path = "./templates/",
  templates_metadata_path = "./templates_metadata/",
  match_grid_metadata_path = "./match_grid_metadata/",
  match_scores_path = "./match_scores/",
  detections_path = "./detections/",
  detection_cuts_path = "./detection_cuts/",
  detection_spectrograms_path = "./detection_spectrograms/",
  validation_outputs_path = "./validation_outputs/",
  validation_diagnostics_path = "./validation_diagnostics/"
)
}
\arguments{
\item{project_path}{The path to the project directory. If NULL, the function
will try to determine the project path automatically based on the path to
the R script that is being executed.}

\item{example_data}{Logical indicating whether to populate the workspace with
example data. Defaults to FALSE.}

\item{app_presets_path}{The path to the app presets directory. This path
contains the species labels file and will be used to store temporary files
generated by the segmentation and validation apps.}

\item{soundscapes_path}{The path to the soundscapes directory. This path is
destined to store soundscape recordings when those are stored within the
project directory.}

\item{soundscapes_metadata_path}{The path to the soundscapes metadata
directory.}

\item{recordings_path}{The path to the recordings directory. This path is
destined to store additional recordings, such as focal recordings, that
from which templates will be extracted.}

\item{roi_tables_path}{The path to the ROI tables directory.}

\item{roi_cuts_path}{The path to the ROI cuts directory. Although the format
is the same as the templates files, the ROI cuts should not be stored
unless explicitly selected as such.}

\item{templates_path}{The path to the templates directory.}

\item{templates_metadata_path}{The path to the templates metadata directory.}

\item{match_grid_metadata_path}{The path to the match grid metadata
directory.}

\item{match_scores_path}{The path to the match scores directory.}

\item{detections_path}{The path to the detections directory.}

\item{detection_cuts_path}{The path to store detection cuts produced by
within the validation app.}

\item{detection_spectrograms_path}{The path to store detection spectrograms
produced by within the validation app.}

\item{validation_outputs_path}{The path to store validated detection tables.}

\item{validation_diagnostics_path}{The path to store validation diagnostics.}
}
\value{
Multiple directories at the project path.
}
\description{
`r lifecycle::badge("experimental")`

  This function sets the working directory to the project path and creates
  the necessary directories if they do not exist. It also creates the species
  labels file if it doesn't exist.
}
\examples{
\dontrun{

# The script below os intended to be run from the a clean R session within a new
# project diorectory. The commands below should not be run in an existing
# project. After running the commands below, close the current R session and
# start a new one within the project directory to run the examples or to build a
# new project from scratch.

# set the workspace to the current directory and populate it with the complete
# example data. Uncomment the command below to run it.

# library(monitoraSom)
# set_workspace(
#     project_path = NULL, ext_soundscapes_path = FALSE, example_data = TRUE
# )

# Or, alternatively, set the workspace to the current directory and populate it
# only with the subdirectories with the recommended structure for a new project.
# When working with soundscapes stored in external directories, which is highly
# recommended, the `ext_soundscapes_path` argument should be set to `TRUE` and
# it will prevent the creation of a soundscapes subdirectory. Uncomment the
# command below to run it.

# library(monitoraSom)
# set_workspace(
#     project_path = NULL, ext_soundscapes_path = TRUE, example_data = FALSE
# )

}
}
