% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_workspace.R
\name{set_workspace}
\alias{set_workspace}
\title{Set the workspace for the monitoraSom package}
\usage{
set_workspace(
  project_path = NULL,
  ext_soundscapes_path = FALSE,
  app_presets_path = "./000_app_presets/",
  soundscapes_path = "./010_soundscapes/",
  soundscapes_metadata_path = "./020_soundscapes_metadata/",
  roi_tables_path = "./030_roi_tables/",
  roi_cuts_path = "./040_roi_cuts/",
  templates_metadata_path = "./050_templates_metadata/",
  match_grid_metadata_path = "./060_match_grid_metadata/",
  match_scores_path = "./070_match_scores/",
  detections_path = "./080_detections/",
  detection_cuts_path = "./090_detection_cuts/",
  detection_spectrograms_path = "./100_detection_spectrograms/",
  validation_outputs_path = "./110_validation_outputs/",
  validation_diagnostics_path = "./120_validation_diagnostics/",
  example_data = FALSE
)
}
\arguments{
\item{project_path}{The path to the project directory. If NULL, the function
will try to determine the project path automatically based on the path to
the R script that is being executed.}

\item{ext_soundscapes_path}{Logical indicating whether the soundscapes path
is external to the project directory. If TRUE, the soundscapes directory
will not be created.}

\item{app_presets_path}{The path to the app presets directory.}

\item{soundscapes_path}{The path to the soundscapes directory.}

\item{soundscapes_metadata_path}{The path to the soundscapes metadata
directory.}

\item{roi_tables_path}{The path to the roi tables directory.}

\item{roi_cuts_path}{The path to the roi cuts directory.}

\item{templates_metadata_path}{The path to the templates metadata directory.}

\item{match_grid_metadata_path}{The path to the match grid metadata directory.}

\item{match_scores_path}{The path to the match scores directory.}

\item{detections_path}{The path to the detections directory.}

\item{detection_cuts_path}{The path to the detection cuts directory.}

\item{detection_spectrograms_path}{The path to the detection spectrograms
directory.}

\item{validation_outputs_path}{The path to the validation outputs directory.}

\item{validation_diagnostics_path}{The path to the validation diagnostics
directory.}

\item{example_data}{Logical indicating whether to populate the workspace with
example data. Defaults to FALSE.}
}
\value{
Multiple directories at the project path.
}
\description{
`r lifecycle::badge("experimental")`

  This function sets the working directory to the project path and creates
  the necessary directories if they do not exist. It also creates the species
  labels file if it doesn't exist.
}
\examples{
\dontrun{

# The script below os intended to be run from the a clean R session within a new
# project diorectory. The commands below should not be run in an existing
# project. After running the commands below, close the current R session and
# start a new one within the project directory to run the examples or to build a
# new project from scratch.

# set the workspace to the current directory and populate it with the complete
# example data. Uncomment the command below to run it.

# library(monitoraSom)
# set_workspace(
#     project_path = NULL, ext_soundscapes_path = FALSE, example_data = TRUE
# )

# Or, alternatively, set the workspace to the current directory and populate it
# only with the subdirectories with the recommended structure for a new project.
# When working with soundscapes stored in external directories, which is highly
# recommended, the `ext_soundscapes_path` argument should be set to `TRUE` and
# it will prevent the creation of a soundscapes subdirectory. Uncomment the
# command below to run it.

# library(monitoraSom)
# set_workspace(
#     project_path = NULL, ext_soundscapes_path = TRUE, example_data = FALSE
# )

}
}
