% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_labels-data.R
\docType{data}
\name{sp_labels}
\alias{sp_labels}
\title{Lists of species for labelling regions of interest}
\format{
## `sp_labels`

A data frame with 7,240 rows and 60 columns:
\describe{
  \item{`CBRO-2021 (Brazil)`}{List of species from Brazil}
  \item{`Aves Argentinas`}{List of species from Argentina}
  ...
}
}
\source{
Todo
}
\usage{
data(sp_lists)
}
\description{
A data frame containing species names to label regions of interest in the
segmentation app.
}
\keyword{datasets}
