% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_matching.R
\name{template_matching}
\alias{template_matching}
\title{Perform template matching on a set of soundscapes using a set of templates}
\usage{
template_matching(
  path_soundscapes,
  recursive_soundscapes = TRUE,
  path_templates,
  recursive_templates = TRUE,
  score_method = "cor",
  output_file = NULL,
  autosave_action = "append",
  buffer_size = "template",
  min_score = NULL,
  min_quant = NULL,
  top_n = NULL,
  ncores = 1
)
}
\arguments{
\item{path_soundscapes}{Path to a directory containing the soundscapes to be
analyzed.}

\item{recursive_soundscapes}{Todo}

\item{path_templates}{Path to a directory containing the templates to be used
for matching.}

\item{recursive_templates}{Todo}

\item{score_method}{The method used for template matching, either "cor" or
"dtw".}

\item{output_file}{Path to the file where the results will be saved.}

\item{autosave_action}{Action to be taken when the user navigates to another
file. Possible values are "append" (default) and "overwrite".}

\item{buffer_size}{The size of the buffer used to calculate the scores for
the template matches, either "template" or "event".}

\item{min_score}{A numeric value between 0 and 0.99 indicating the minimum
score of the detections that will be kept. Defaults to NULL, which returns
all available detections.}

\item{min_quant}{A numeric value between 0 and 1 indicating the minimum score
quantile of the kept detections. Defaults to NULL, which returns all
available detections.}

\item{top_n}{An integer indicating the maximum number of peaks to be
returned, selected according to the highest scores available. Defaults to
NULL, which return all available detections. It should be noted that
because the peak quantiles are callculated within each score vector, the
top_n parameter is applied to each score vector separately, and not to the
whole matching grid.}

\item{ncores}{An integer indicating the number of cores to be used for
parallelization. Default is 1.}
}
\value{
A dataframe with the detected events.
}
\description{
`r lifecycle::badge("experimental")`

  This function performs template matching on a set of soundscapes using a
  set of templates, and returns a dataframe with the detected events.
}
