% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_matching.R
\name{template_matching}
\alias{template_matching}
\title{Perform template matching on a set of soundscapes using a set of templates}
\usage{
template_matching(
  path_soundscapes,
  path_templates,
  template_type = "standalone",
  score_method = "cor",
  buffer_size = "template",
  ncores = 1
)
}
\arguments{
\item{path_soundscapes}{Path to a directory containing the soundscapes to be
analyzed.}

\item{path_templates}{Path to a directory containing the templates to be used
for matching.}

\item{template_type}{The type of templates, either "standalone" or "library".}

\item{score_method}{The method used for template matching, either "cor" or
"dtw".}

\item{buffer_size}{The size of the buffer used to calculate the scores for
the template matches, either "template" or "event".}

\item{ncores}{The number of cores to be used for parallel processing.}
}
\value{
A dataframe with the detected events.
}
\description{
`r lifecycle::badge("experimental")`

This function performs template matching on a set of soundscapes using a set
of templates, and returns a dataframe with the detected events.
}
\examples{
df_detections <- template_matching(
  path_soundscapes = "soundscapes/",
  path_templates = "templates/",
  template_type = "standalone",
  score_method = "cor",
  buffer_size = "template",
  ncores = 4
)
}
