% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_by_overlap.R
\name{validate_by_overlap}
\alias{validate_by_overlap}
\title{Batch validation of detections by overlap with Regions of Interest (ROIs)}
\usage{
validate_by_overlap(
  df_detecs,
  df_rois,
  validation_user = NULL,
  recursive = FALSE,
  output_path = NULL
)
}
\arguments{
\item{df_detecs}{A data frame containing detections as in the output of
`fetch_score_peaks_n()`. Alternatively, a path to a CSV file
containing the detections (it accepts only one CSV file at a time).}

\item{df_rois}{A data frame containing Regions of Interest (ROIs) as in the
output of `fetch_rois_n()`. Alternatively, a path to a directory containing
the ROI tables as CSV files can be provided.}

\item{validation_user}{A character string specifying the user name.}

\item{recursive}{A logical value indicating whether to search for ROIs
recursively when a path is provided instead of a data frame. Defaults to
FALSE.}

\item{output_path}{A character string specifying the path to the output file.
Defaults to NULL. it is recommended to export it as a CSV file to the
"validation_output/" directory to avoid replacement of existing detection
files.}
}
\value{
A data frame with the validated detections.
}
\description{
`r lifecycle::badge("experimental")`

  This function takes in two data frames, one containing Regions of Interest
  (ROIs) and another containing detections. It then validates the detections
  by checking for overlaps with the ROIs. The function returns a data frame
  with the validated detections.
}
